Add files via upload
Documentation:

We split our code into email, process_a, and process_b. For this program, we did not use a main. In the email go file, we imported the functions that were going to be used throughout the programs and then created an email structure that contained the data types of the five fields that we needed to define (To, From, Title, Date, and Content). For the process_a file, we imported six packages. We decided that process_a would be the client in this case because it sent the message to process_b. In the code we used CONNECT as a way for the TCP Client to connect to the TCP server, which is process_b. We put a section which is a for loop that allows the user to input information. In the process_b file, we also imported the same six packages. Those six packages were used to create the TCP server. Then we used accept as a way for the TCP server and the TCP client to interact with one another.